requirements.txt
pywebview>=5.0.0

